xerver pixman libX11 mesa git update Jun 13 2021
xserver          commit 7e7c147105ac704a7b33bdee0f85dc9ccdd8f747
libX11           commit f906fe8e9769e4313294b68e61c402610ade69da
pixman           commit 4251202d9dbb7503976fa6c651747b8e99ea0988
mesa             commit a1c56b80915a1105c5a62aa6ff3ac71c1edd143d

+++ mesalib/.gitlab-ci/container/lava_build.sh
@@ -3,6 +3,8 @@
 set -e
 set -o xtrace
 
+export DEBIAN_FRONTEND=noninteractive
+
 check_minio()
 {
     MINIO_PATH="${MINIO_HOST}/mesa-lava/$1/${DISTRIBUTION_TAG}/${DEBIAN_ARCH}"
@@ -65,49 +67,43 @@ if [[ -e /cross_file-$DEBIAN_ARCH.txt ]]; then
 fi
 
 apt-get update
-apt-get install -y automake \
+apt-get install -y --no-remove \
+                   automake \
                    bc \
                    cmake \
                    debootstrap \
                    git \
-                   libboost-dev \
+                   glslang-tools \
                    libegl1-mesa-dev \
                    libgbm-dev \
                    libgles2-mesa-dev \
-                   libpcre3-dev \
-                   libpng-dev \
-                   libpython3-dev \
                    libssl-dev \
+                   libudev-dev \
                    libvulkan-dev \
                    libwaffle-dev \
-                   libxcb-keysyms1-dev \
+                   libwayland-dev \
+                   libx11-xcb-dev \
                    libxkbcommon-dev \
                    patch \
-                   python3-dev \
                    python3-distutils \
                    python3-mako \
                    python3-numpy \
                    python3-serial \
-                   qt5-default \
-                   qt5-qmake \
-                   qtbase5-dev \
                    wget
 
 
 if [[ "$DEBIAN_ARCH" = "armhf" ]]; then
-    apt-get install -y libboost-dev:armhf \
+    apt-get install -y --no-remove \
                        libegl1-mesa-dev:armhf \
                        libelf-dev:armhf \
                        libgbm-dev:armhf \
                        libgles2-mesa-dev:armhf \
-                       libpcre3-dev:armhf \
-                       libpng-dev:armhf \
-                       libpython3-dev:armhf \
+                       libudev-dev:armhf \
                        libvulkan-dev:armhf \
                        libwaffle-dev:armhf \
-                       libxcb-keysyms1-dev:armhf \
-                       libxkbcommon-dev:armhf \
-                       qtbase5-dev:armhf
+                       libwayland-dev:armhf \
+                       libx11-xcb-dev:armhf \
+                       libxkbcommon-dev:armhf
 fi
 
 
@@ -130,29 +126,8 @@ mv /deqp /lava-files/rootfs-${DEBIAN_ARCH}/.
 
 
 ############### Build piglit
-if [ -n "$INCLUDE_PIGLIT" ]; then
-    . .gitlab-ci/container/build-piglit.sh
-    mv /piglit /lava-files/rootfs-${DEBIAN_ARCH}/.
-fi
-
-
-############### Build apitrace
-. .gitlab-ci/container/build-apitrace.sh
-mkdir -p /lava-files/rootfs-${DEBIAN_ARCH}/apitrace
-mv /apitrace/build /lava-files/rootfs-${DEBIAN_ARCH}/apitrace
-rm -rf /apitrace
-
-mkdir -p /lava-files/rootfs-${DEBIAN_ARCH}/waffle
-mv /waffle/build /lava-files/rootfs-${DEBIAN_ARCH}/waffle
-rm -rf /waffle
-
-
-############### Build renderdoc
-EXTRA_CMAKE_ARGS+=" -DENABLE_XCB=false"
-. .gitlab-ci/container/build-renderdoc.sh
-mkdir -p /lava-files/rootfs-${DEBIAN_ARCH}/renderdoc
-mv /renderdoc/build /lava-files/rootfs-${DEBIAN_ARCH}/renderdoc
-rm -rf /renderdoc
+. .gitlab-ci/container/build-piglit.sh
+mv /piglit /lava-files/rootfs-${DEBIAN_ARCH}/.
 
 
 ############### Build libdrm
@@ -162,7 +137,7 @@ EXTRA_MESON_ARGS+=" -D prefix=/libdrm"
 
 ############### Cross-build kernel
 mkdir -p kernel
-wget -qO- ${KERNEL_URL} | tar -xz --strip-components=1 -C kernel
+wget -qO- ${KERNEL_URL} | tar -xj --strip-components=1 -C kernel
 pushd kernel
 
 # The kernel doesn't like the gold linker (or the old lld in our debians).
@@ -175,11 +150,7 @@ for i in /usr/bin/*-ld /usr/bin/ld; do
 done
 export PATH=`pwd`/ld-links:$PATH
 
-if [ -n "$INSTALL_KERNEL_MODULES" ]; then
-    # Disable all modules in defconfig, so we only build the ones we want
-    sed -i 's/=m/=n/g' ${DEFCONFIG}
-fi
-
+export LOCALVERSION="`basename $KERNEL_URL`"
 ./scripts/kconfig/merge_config.sh ${DEFCONFIG} ../.gitlab-ci/container/${KERNEL_ARCH}.config
 make ${KERNEL_IMAGE_NAME}
 for image in ${KERNEL_IMAGE_NAME}; do
@@ -191,12 +162,12 @@ if [[ -n ${DEVICE_TREES} ]]; then
     cp ${DEVICE_TREES} /lava-files/.
 fi
 
-if [ -n "$INSTALL_KERNEL_MODULES" ]; then
+if [[ ${DEBIAN_ARCH} = "amd64" ]]; then
     make modules
     INSTALL_MOD_PATH=/lava-files/rootfs-${DEBIAN_ARCH}/ make modules_install
 fi
 
-if [[ ${DEBIAN_ARCH} = "arm64" ]] && which mkimage > /dev/null; then
+if [[ ${DEBIAN_ARCH} = "arm64" ]]; then
     make Image.lzma
     mkimage \
         -f auto \
@@ -206,33 +177,32 @@ if [[ ${DEBIAN_ARCH} = "arm64" ]] && which mkimage > /dev/null; then
         -C lzma\
         -b arch/arm64/boot/dts/qcom/sdm845-cheza-r3.dtb \
         /lava-files/cheza-kernel
+    KERNEL_IMAGE_NAME+=" cheza-kernel"
 fi
 
 popd
 rm -rf kernel
 
 ############### Delete rust, since the tests won't be compiling anything.
-rm -rf /root/.rustup /root/.cargo
+rm -rf /root/.cargo
 
 ############### Create rootfs
 set +e
-debootstrap \
-    --variant=minbase \
-    --arch=${DEBIAN_ARCH} \
+if ! debootstrap \
+     --variant=minbase \
+     --arch=${DEBIAN_ARCH} \
      --components main,contrib,non-free \
-    buster \
-    /lava-files/rootfs-${DEBIAN_ARCH}/ \
-    http://deb.debian.org/debian
-
-cat /lava-files/rootfs-${DEBIAN_ARCH}/debootstrap/debootstrap.log
+     bullseye \
+     /lava-files/rootfs-${DEBIAN_ARCH}/ \
+     http://deb.debian.org/debian; then
+    cat /lava-files/rootfs-${DEBIAN_ARCH}/debootstrap/debootstrap.log
+    exit 1
+fi
 set -e
 
 cp .gitlab-ci/container/create-rootfs.sh /lava-files/rootfs-${DEBIAN_ARCH}/.
-cp .gitlab-ci/container/llvm-snapshot.gpg.key /lava-files/rootfs-${DEBIAN_ARCH}/.
-chroot /lava-files/rootfs-${DEBIAN_ARCH} \
-    sh -c "INCLUDE_PIGLIT=$INCLUDE_PIGLIT sh /create-rootfs.sh"
+chroot /lava-files/rootfs-${DEBIAN_ARCH} sh /create-rootfs.sh
 rm /lava-files/rootfs-${DEBIAN_ARCH}/create-rootfs.sh
-rm /lava-files/rootfs-${DEBIAN_ARCH}/llvm-snapshot.gpg.key
 
 
 ############### Install the built libdrm
@@ -244,63 +214,32 @@ find /libdrm/ -name lib\*\.so\* | xargs cp -t /lava-files/rootfs-${DEBIAN_ARCH}/
 rm -rf /libdrm
 
 
+if [ ${DEBIAN_ARCH} = arm64 ]; then
+    # Make a gzipped copy of the Image for db410c.
+    gzip -k /lava-files/Image
+    KERNEL_IMAGE_NAME+=" Image.gz"
+fi
+
 du -ah /lava-files/rootfs-${DEBIAN_ARCH} | sort -h | tail -100
 pushd /lava-files/rootfs-${DEBIAN_ARCH}
   tar czf /lava-files/lava-rootfs.tgz .
 popd
 
-if [ ${DEBIAN_ARCH} = arm64 ]; then
-    # Pull down a specific build of qcomlt/release/qcomlt-5.4 8c79b3d12355
-    # ("Merge tag 'v5.4.23' into release/qcomlt-5.4"), where I used the
-    # .config from
-    # http://snapshots.linaro.org/96boards/dragonboard820c/linaro/debian/457/config-5.4.0-qcomlt-arm64
-    # with the following merged in:
-    #
-    # CONFIG_DRM=y
-    # CONFIG_DRM_MSM=y
-    # CONFIG_ATL1C=y
-    #
-    # Reason: 5.5 has a big stack of oopses and warns on db820c.  4.14-5.4
-    # linaro kernel binaries (see above .config link) have these as modules
-    # and distributed the modules only in the debian system, not the initrd,
-    # so they're very hard to extract (involving simg2img and loopback
-    # mounting).  4.11 is missing d72fea538fe6 ("drm/msm: Fix the check for
-    # the command size") so it can't actually run fredreno.  qcomlt-4.14 is
-    # unstable at boot (~10% instaboot rate).  The 5.4 qcomlt kernel with msm
-    # built in seems like the easiest way to go.
-    wget https://people.freedesktop.org/~anholt/qcomlt-5.4-msm-build/Image.gz -O Image.gz \
-         -O /lava-files/db820c-kernel
-    wget https://people.freedesktop.org/~anholt/qcomlt-5.4-msm-build/apq8096-db820c.dtb \
-         -O /lava-files/db820c.dtb
-
-    # Make a gzipped copy of the Image for db410c.
-    gzip -k /lava-files/Image
-
-    # Add missing a630 firmware, added to debian packge in apr 2020
-    wget https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/plain/qcom/a630_gmu.bin \
-         -O /lava-files/rootfs-arm64/lib/firmware/qcom/a630_gmu.bin
-    wget https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/plain/qcom/a630_sqe.fw \
-         -O /lava-files/rootfs-arm64/lib/firmware/qcom/a630_sqe.fw
-fi
-
 . .gitlab-ci/container/container_post_build.sh
 
 ############### Upload the files!
-if [ -n "$UPLOAD_FOR_LAVA" ]; then
-    ci-fairy minio login $CI_JOB_JWT
-    FILES_TO_UPLOAD="lava-rootfs.tgz \
-                     $KERNEL_IMAGE_NAME"
+ci-fairy minio login $CI_JOB_JWT
+FILES_TO_UPLOAD="lava-rootfs.tgz \
+                 $KERNEL_IMAGE_NAME"
 
-    if [[ -n $DEVICE_TREES ]]; then
-        FILES_TO_UPLOAD="$FILES_TO_UPLOAD $(basename -a $DEVICE_TREES)"
-    fi
-
-    for f in $FILES_TO_UPLOAD; do
-        ci-fairy minio cp /lava-files/$f \
-            minio://${MINIO_PATH}/$f
-    done
-
-    touch /lava-files/done
-    ci-fairy minio cp /lava-files/done minio://${MINIO_PATH}/done
+if [[ -n $DEVICE_TREES ]]; then
+    FILES_TO_UPLOAD="$FILES_TO_UPLOAD $(basename -a $DEVICE_TREES)"
 fi
 
+for f in $FILES_TO_UPLOAD; do
+    ci-fairy minio cp /lava-files/$f \
+             minio://${MINIO_PATH}/$f
+done
+
+touch /lava-files/done
+ci-fairy minio cp /lava-files/done minio://${MINIO_PATH}/done
